Refactor forum data types and session management - Breaking Schema Change
Updated various timestamp and identifier columns from int to long to be able to store DateTime as Tick (long integer) or varchar(128) to handle larger values. Added ForumSessionHelper and IpAddressHelper for session and IP management. Introduced ForumSessionService and ForumSessionService, replacing SessionService and ISessionService. Updated Index.cshtml.cs and Program.cs to use new session management.

+++ source/digioz.Forum/digioz.Forum/Models/ForumAttachment.cs
@@ -31,7 +31,7 @@ public partial class ForumAttachment
 
     public int FileSize { get; set; }
 
-    public int FileTime { get; set; }
+    public long FileTime { get; set; }
 
     public byte Thumbnail { get; set; }
 }
