+++ source/digioz.Forum/digioz.Forum/Models/ForumUser.cs
@@ -29,13 +29,13 @@ public partial class ForumUser
 
     public string UserBirthday { get; set; } = null!;
 
-    public int UserLastVisit { get; set; }
+    public long UserLastVisit { get; set; }
 
-    public int UserLastActive { get; set; }
+    public long UserLastActive { get; set; }
 
-    public int UserLastMark { get; set; }
+    public long UserLastMark { get; set; }
 
-    public int UserLastPostTime { get; set; }
+    public long UserLastPostTime { get; set; }
 
     public string UserLastPage { get; set; } = null!;
 
@@ -45,13 +45,13 @@ public partial class ForumUser
 
     public byte UserWarnings { get; set; }
 
-    public int UserLastWarning { get; set; }
+    public long UserLastWarning { get; set; }
 
     public byte UserLoginAttempts { get; set; }
 
     public byte UserInactiveReason { get; set; }
 
-    public int UserInactiveTime { get; set; }
+    public long UserInactiveTime { get; set; }
 
     public long UserPosts { get; set; }
 
@@ -71,13 +71,13 @@ public partial class ForumUser
 
     public int UserUnreadPrivmsg { get; set; }
 
-    public int UserLastPrivmsg { get; set; }
+    public long UserLastPrivmsg { get; set; }
 
     public byte UserMessageRules { get; set; }
 
     public int UserFullFolder { get; set; }
 
-    public int UserEmailTime { get; set; }
+    public long UserEmailTime { get; set; }
 
     public short UserTopicShowDays { get; set; }
 
@@ -125,11 +125,11 @@ public partial class ForumUser
 
     public string UserActkey { get; set; } = null!;
 
-    public int UserActkeyExpiration { get; set; }
+    public long UserActkeyExpiration { get; set; }
 
     public string ResetToken { get; set; } = null!;
 
-    public int ResetTokenExpiration { get; set; }
+    public long ResetTokenExpiration { get; set; }
 
     public string UserNewpasswd { get; set; } = null!;
 
@@ -139,5 +139,5 @@ public partial class ForumUser
 
     public byte UserReminded { get; set; }
 
-    public int UserRemindedTime { get; set; }
+    public long UserRemindedTime { get; set; }
 }
